+++ packages/mg-components/.storybook/main.ts
@@ -121,6 +121,7 @@ const config: StorybookConfig = {
         'v5.27.2': 'https://626149b307606d003ada26b4-rgtwzbiszp.chromatic.com',
         'v5.27.3': 'https://626149b307606d003ada26b4-nxqzxvfuib.chromatic.com',
         'v5.27.4': 'https://626149b307606d003ada26b4-pponmjenit.chromatic.com',
+        'v5.28.0': 'https://626149b307606d003ada26b4-grmjndcjgx.chromatic.com',
       },
     },
   },