+++ apps/angularjs/CHANGELOG.md
@@ -1,5 +1,22 @@
 # angularjs
 
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies [6195e7a]
+  - @mgdis/notification-center@1.1.14
+  - @mgdis/mg-components@5.28.1
+
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [4ee49fe]
+  - @mgdis/mg-components-helpers@4.1.6
+  - @mgdis/notification-center@1.1.13
+  - @mgdis/mg-components@5.28.1
+
 ## 0.0.24
 
 ### Patch Changes